travis: Switch to dpdk main branch.
Make this branch point to current DPDK main branch so that we can
track API breakage.

Also the default branch name in DPDK is changed from master to main. This
patch reflects the same on travis builds for dpdk-latest branch.

Note: this should not be merged to master, intended for dpdk-latest
branch only.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Sunil Pai G <sunil.pai.g@intel.com>
Co-authored-by: Sunil Pai G <sunil.pai.g@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>